+++ tests/test_minimal/test_utils/test_get_json_schema_from_method_signature.py
@@ -213,7 +213,7 @@ def test_get_json_schema_from_example_data_interface():
                 "type": "string",
                 "description": "Path to the folder of .mpx files.",
             },
-            "verbose": {"default": True, "type": "boolean", "description": "Allows verbose.\nDefault is True."},
+            "verbose": {"default": False, "type": "boolean", "description": "Allows verbose.\nDefault is False."},
             "es_key": {"default": "ElectricalSeries", "type": "string"},
         },
         "required": ["folder_path"],