+++ src/neuroconv/datainterfaces/behavior/lightningpose/lightningposedatainterface.py
@@ -65,7 +65,7 @@ class LightningPoseDataInterface(BaseTemporalAlignmentInterface):
         file_path: FilePath,
         original_video_file_path: FilePath,
         labeled_video_file_path: Optional[FilePath] = None,
-        verbose: bool = True,
+        verbose: bool = False,
     ):
         """
         Interface for writing pose estimation data from the Lightning Pose algorithm.
@@ -78,7 +78,7 @@ class LightningPoseDataInterface(BaseTemporalAlignmentInterface):
             Path to the original video file (.mp4).
         labeled_video_file_path : a string or a path, optional
             Path to the labeled video file (.mp4).
-        verbose : bool, default: True
+        verbose : bool, default: False
             controls verbosity. ``True`` by default.
         """